+++ apps/cowswap-frontend/src/modules/analytics/events.ts
@@ -25,6 +25,7 @@ export enum Category {
   COW_FORTUNE = 'CoWFortune',
   SURPLUS_MODAL = 'Surplus Modal',
   PROGRESS_BAR = 'Progress Bar',
+  NOTIFICATIONS = 'Notifications',
 }
 
 export function shareFortuneTwitterAnalytics() {
@@ -320,5 +321,14 @@ export function shareSurplusOnTwitter() {
   cowAnalytics.sendEvent({
     category: Category.SURPLUS_MODAL,
     action: `Share on Twitter`,
+  })
+}
+
+export function clickNotifications(event: string, notificationId?: number, title?: string) {
+  cowAnalytics.sendEvent({
+    category: Category.NOTIFICATIONS,
+    action: event,
+    value: notificationId,
+    label: title,
   })
 }