+++ apps/cowswap-frontend/src/modules/notifications/containers/NotificationBell.tsx
@@ -7,7 +7,6 @@ import { UI } from '@cowprotocol/ui'
 import SVG from 'react-inlinesvg'
 import styled from 'styled-components/macro'
 
-import { useUnreadNotifications } from '../hooks/useUnreadNotifications'
 
 const Icon = styled.div<{ hasNotification?: boolean }>`
   --size: 18px;
@@ -59,14 +58,12 @@ const Icon = styled.div<{ hasNotification?: boolean }>`
 
 interface NotificationBellProps {
   onClick: Command
+  unreadCount: number
 }
 
-export function NotificationBell({ onClick }: NotificationBellProps) {
-  const unreadNotifications = useUnreadNotifications()
-  const unreadNotificationsCount = Object.keys(unreadNotifications).length
-
+export function NotificationBell({ onClick, unreadCount }: NotificationBellProps) {
   return (
-    <Icon hasNotification={unreadNotificationsCount > 0} onClick={onClick}>
+    <Icon hasNotification={unreadCount > 0} onClick={onClick}>
       <SVG src={ICON_NOTIFICATION} />
     </Icon>
   )