+++ libs/common-utils/src/uriToHttp.ts
@@ -17,15 +17,12 @@ export function uriToHttp(uri: string): string[] {
     case 'http':
       return ['https' + uri.substr(4), uri]
     case 'ipfs':
-      // eslint-disable-next-line no-case-declarations
       const hash = uri.match(/^ipfs:(\/\/)?(ipfs\/)?(.*)$/i)?.[3] // TODO: probably a bug on original code
-      return [`https://cloudflare-ipfs.com/ipfs/${hash}/`, `https://ipfs.io/ipfs/${hash}/`]
+      return [`https://ipfs.io/ipfs/${hash}/`]
     case 'ipns':
-      // eslint-disable-next-line no-case-declarations
       const name = uri.match(/^ipns:(\/\/)?(.*)$/i)?.[2]
-      return [`https://cloudflare-ipfs.com/ipns/${name}/`, `https://ipfs.io/ipns/${name}/`]
+      return [`https://ipfs.io/ipns/${name}/`]
     case 'ar':
-      // eslint-disable-next-line no-case-declarations
       const tx = uri.match(/^ar:(\/\/)?(.*)$/i)?.[2]
       return [`https://arweave.net/${tx}`]
     default: