+++ libs/core/src/cms/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useCmsSolversInfo'
+export * from './useSolversInfo'